# target/riscv: refactor read_memory_progbuf()