@en-sc 2023-07-07
> How does this change affect the performance when reading a lot of data?

From my testing, there is no measurable difference, since at the core the approach is the same. The change affects busy response handling, which is a rare occurrence.

The performance was tested by dumping an image (on spike) as follows:
```
openocd -f ... -c 'init' -c 'reset halt' -c 'dump_image dump.img 0x80000000 [expr 4*1024*1024]' -c 'shutdown'
```
Before the change:
```
Open On-Chip Debugger 0.12.0+dev-00573-g92c031926 (2023-07-07-13:53)
...
dumped 4194304 bytes in 60.885689s (67.274 KiB/s)
...
```
After the change:
```
Open On-Chip Debugger 0.12.0+dev-00575-g291433c9f (2023-07-07-13:51)
...
Warn : [riscv.cpu] re-reading memory from addresses 0x80000008 and 0x80000010
dumped 4194304 bytes in 55.689625s (73.551 KiB/s)
...
```

There seems to be a speed-up, but it is not reproduced on hardware - I have observed a similarly small decrease in read speed.